pyproject.toml
[tool.poetry]
name = "djangotemplate"
version = "0.1.0"
description = ""
authors = ["pchenr1que1 <pchenr1que1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"
django-extensions = "^3.2.1"
djangorestframework-simplejwt = "^5.2.2"
logging = "^0.4.9.6"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
